Move editor to separate folder
